Replace previous graphql module with next
