--- aquadoggo/src/dynamic_graphql/schema.rs
@@ -1,327 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-//! Build and manage a GraphQL schema including dynamic parts of the schema.
-use std::sync::Arc;
-
-use async_graphql::dynamic::{Object, Schema};
-use async_graphql::{Request, Response};
-use dynamic_graphql::internal::Registry;
-use log::{debug, info};
-use p2panda_rs::Human;
-use tokio::sync::Mutex;
-
-use crate::bus::ServiceSender;
-use crate::db::SqlStore;
-use crate::dynamic_graphql::mutations::{MutationRoot, Publish};
-use crate::dynamic_graphql::scalars::{
-    DocumentIdScalar, DocumentViewIdScalar, EncodedEntryScalar, EncodedOperationScalar,
-    EntryHashScalar, LogIdScalar, PublicKeyScalar, SeqNumScalar,
-};
-use crate::dynamic_graphql::schema_builders::{
-    build_document_schema, build_document_field_schema, build_document_query, build_next_args_query,
-};
-use crate::dynamic_graphql::types::{DocumentMeta, NextArguments};
-use crate::dynamic_graphql::utils::fields_name;
-use crate::schema::SchemaProvider;
-
-/// Returns GraphQL API schema for p2panda node.
-///
-/// Builds the root schema that can handle all GraphQL requests from clients (Client API) or other
-/// nodes (Node API).
-pub async fn build_root_schema(
-    store: SqlStore,
-    tx: ServiceSender,
-    schema_provider: SchemaProvider,
-) -> Schema {
-    // Get all schema from the schema provider.
-    let all_schema = schema_provider.all().await;
-
-    // Using dynamic-graphql we create a registry and add types.
-    let registry = Registry::new()
-        .register::<NextArguments>()
-        .register::<MutationRoot>()
-        .register::<Publish>()
-        .register::<DocumentIdScalar>()
-        .register::<DocumentMeta>()
-        .register::<DocumentViewIdScalar>()
-        .register::<EncodedEntryScalar>()
-        .register::<EncodedOperationScalar>()
-        .register::<EntryHashScalar>()
-        .register::<LogIdScalar>()
-        .register::<PublicKeyScalar>()
-        .register::<SeqNumScalar>();
-
-    // Construct the schema builder.
-    let mut schema_builder = Schema::build("Query", Some("MutationRoot"), None);
-
-    // Populate it with the registered types. We can now use these in any following dynamically
-    // created query object fields.
-    schema_builder = registry.apply_into_schema_builder(schema_builder);
-
-    // Construct the root query object.
-    let mut root_query = Object::new("Query");
-
-    // Loop through all schema retrieved from the schema store, create types and a root query for the
-    // documents they describe.
-    for schema in all_schema {
-        // Construct the document fields object which will be named `<schema_id>Field`.
-        let schema_field_name = fields_name(&schema.id().to_string());
-        let mut document_schema_fields = Object::new(&schema_field_name);
-
-        // For every field in the schema we create a type with a resolver.
-        //
-        // TODO: We can optimize the field resolution methods later with a data loader.
-        for (name, field_type) in schema.fields() {
-            document_schema_fields = build_document_field_schema(document_schema_fields, name, field_type);
-        }
-
-        // Construct the document schema which has "fields" and "meta" fields.
-        let document_schema = build_document_schema(&schema);
-
-        // Register a schema and schema fields type for every schema.
-        schema_builder = schema_builder.register(document_schema_fields).register(document_schema);
-
-        // Add a query object for each schema. It offers an interface to retrieve a single
-        // document of this schema by it's document id or view id. It's resolver parses and
-        // validates the passed parameters, then forwards them up to the children query fields.
-        root_query = build_document_query(root_query, &schema);
-    }
-
-    // Add next args to the query object.
-    let root_query = build_next_args_query(root_query);
-
-    // Build the schema.
-    schema_builder
-        .register(root_query)
-        .data(store)
-        .data(schema_provider)
-        .data(tx)
-        .finish()
-        .unwrap()
-}
-
-/// List of created GraphQL root schemas.
-type GraphQLSchemas = Arc<Mutex<Vec<Schema>>>;
-
-/// Shared types between GraphQL schemas.
-#[derive(Clone, Debug)]
-pub struct GraphQLSharedData {
-    /// Database interface.
-    store: SqlStore,
-
-    /// Communication bus interface to send messages to other services.
-    tx: ServiceSender,
-
-    /// Schema provider giving us access to currently known schemas.
-    schema_provider: SchemaProvider,
-}
-
-/// Builds new GraphQL schemas dynamically and executes the latest GraphQL schema for incoming
-/// queries.
-///
-/// This manager allows us to introduce new GraphQL schemas during runtime as it internally handles
-/// a list of schemas (behind a mutex) and automatically picks the "latest" as soon as a query
-/// needs to be executed.
-///
-/// With this we can easily add "new" schemas to the list in the background while current queries
-/// still get processed using the "old" schema.
-//
-// @TODO: This manager does not "clean up" outdated schemas yet, they will just be appended to
-// an ever-growing list.
-//
-// WARNING: As soon as we start implementing GraphQL schema clean-up, we need to make sure to also
-// free the used memory for all leaked schema data we've created. Otherwise this will lead to a
-// memory leak! See `static_provider` module for more information (and useful tools) on this whole
-// topic.
-#[derive(Clone)]
-pub struct GraphQLSchemaManager {
-    /// List of all built GraphQL root schemas.
-    schemas: GraphQLSchemas,
-
-    /// Commonly shared types for GraphQL schemas.
-    shared: GraphQLSharedData,
-}
-
-impl GraphQLSchemaManager {
-    /// Returns a new instance of `GraphQLSchemaManager`.
-    pub async fn new(store: SqlStore, tx: ServiceSender, schema_provider: SchemaProvider) -> Self {
-        let schemas = Arc::new(Mutex::new(Vec::new()));
-        let shared = GraphQLSharedData {
-            store,
-            tx,
-            schema_provider,
-        };
-
-        // Create manager instance and spawn internal watch task
-        let manager = Self { schemas, shared };
-        manager.spawn_schema_added_task().await;
-
-        manager
-    }
-
-    /// Subscribes to `SchemaProvider` for newly added schemas.
-    ///
-    /// This spawns a task which listens to new p2panda schemas to accordingly build a GraphQL
-    /// schema which will be added to the list.
-    async fn spawn_schema_added_task(&self) {
-        let shared = self.shared.clone();
-        let schemas = self.schemas.clone();
-
-        info!("Subscribing GraphQL manager to schema provider");
-        let mut on_schema_added = shared.schema_provider.on_schema_added();
-
-        // Create the new GraphQL based on the current state of known p2panda application schemas
-        async fn rebuild(shared: GraphQLSharedData, schemas: GraphQLSchemas) {
-            let schema = build_root_schema(shared.store, shared.tx, shared.schema_provider).await;
-            schemas.lock().await.push(schema);
-        }
-
-        // Always build a schema right at the beginning as we don't have one yet
-        rebuild(shared.clone(), schemas.clone()).await;
-        debug!("Finished building initial GraphQL schema");
-
-        // Spawn a task which reacts to newly registered p2panda schemas
-        tokio::task::spawn(async move {
-            loop {
-                match on_schema_added.recv().await {
-                    Ok(schema_id) => {
-                        info!(
-                            "Changed schema {}, rebuilding GraphQL API",
-                            schema_id.display()
-                        );
-                        rebuild(shared.clone(), schemas.clone()).await;
-                    }
-                    Err(err) => {
-                        panic!("Failed receiving schema updates: {}", err)
-                    }
-                }
-            }
-        });
-    }
-
-    /// Executes an incoming GraphQL query.
-    ///
-    /// This method makes sure the GraphQL query will be executed by the latest given schema the
-    /// manager knows about.
-    pub async fn execute(&self, request: impl Into<Request>) -> Response {
-        self.schemas
-            .lock()
-            .await
-            .last()
-            .expect("No schema given yet")
-            .execute(request)
-            .await
-    }
-}
-
-impl std::fmt::Debug for GraphQLSchemaManager {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // `schemas` does not implement `Debug` but we can at least print the other fields
-        f.debug_struct("GraphQLSchemaManager")
-            .field("shared", &self.shared)
-            .finish()
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use async_graphql::{value, Response};
-    use p2panda_rs::schema::FieldType;
-    use p2panda_rs::test_utils::constants::PRIVATE_KEY;
-    use p2panda_rs::test_utils::fixtures::key_pair;
-    use rstest::rstest;
-    use serde_json::{json, Value};
-    use serial_test::serial;
-
-    use crate::test_utils::{add_schema, graphql_test_client, test_runner, TestNode};
-
-    #[rstest]
-    // Note: This test uses the underlying static schema provider which is a static mutable data
-    // store, accessible across all test runner threads in parallel mode. To prevent overwriting
-    // data across threads we have to run this test in serial.
-    //
-    // Read more: https://users.rust-lang.org/t/static-mutables-in-tests/49321
-    #[serial]
-    fn schema_updates() {
-        test_runner(|mut node: TestNode| async move {
-            // Create test client in the beginning so it is initialised with just the system
-            // schemas. Then we create a new application schema to test that the graphql schema
-            // is updated and we can query the changed schema.
-            let client = graphql_test_client(&node).await;
-
-            // This test uses a fixed private key to allow us to anticipate the schema typename.
-            let key_pair = key_pair(PRIVATE_KEY);
-            let type_name =
-                "schema_name_0020f71816cb258a24bab65d63d42298467cc983670adc0bb2dd35e19f937011fbb2";
-
-            // Check that the schema does not exist yet.
-            let response = client
-                .post("/graphql")
-                .json(&json!({
-                    "query": format!(
-                        r#"{{
-                        schema: __type(name: "{}") {{
-                            name,
-                        }},
-                    }}"#,
-                        type_name,
-                    ),
-                }))
-                .send()
-                .await;
-            let response: Response = response.json().await;
-
-            assert_eq!(
-                response.data,
-                value!({
-                    "schema": Value::Null,
-                }),
-                "\n{:#?}\n",
-                response.errors
-            );
-
-            // Add schema to node.
-            let schema = add_schema(
-                &mut node,
-                "schema_name",
-                vec![("bool_field", FieldType::Boolean)],
-                &key_pair,
-            )
-            .await;
-
-            assert_eq!(
-                schema.id().to_string(),
-                type_name,
-                "Please update `type_name` const above to fix this test."
-            );
-
-            // Query gql schema.
-            let response = client
-                .post("/graphql")
-                .json(&json!({
-                    "query": format!(
-                        r#"{{
-                        schema: __type(name: "{}") {{
-                            name,
-                        }},
-                    }}"#,
-                        type_name,
-                    ),
-                }))
-                .send()
-                .await;
-            let response: Response = response.json().await;
-
-            assert_eq!(
-                response.data,
-                value!({
-                    "schema": {
-                        "name": type_name
-                    },
-                }),
-                "\n{:#?}\n",
-                response.errors
-            );
-        });
-    }
-}
