--- aquadoggo/src/graphql/client/static_types/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod next_arguments;
-
-pub use next_arguments::NextArguments;
