--- aquadoggo/src/graphql/replication/mod.rs
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-//! API for replicating data with other p2panda nodes
-pub mod client;
-mod query;
-mod response;
-
-pub use client::entries_newer_than_seq_num;
-pub use query::ReplicationRoot;
-pub use response::EncodedEntryAndOperation;
