+++ aquadoggo/src/graphql/scalars/encoded_entry_scalar.rs
@@ -1,24 +1,24 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 use anyhow::anyhow;
-use async_graphql::{InputValueError, InputValueResult, Scalar, ScalarType, Value};
+use dynamic_graphql::{Scalar, ScalarValue, Value, Result, Error};
 use p2panda_rs::entry::EncodedEntry;
 use serde::{Deserialize, Serialize};
 
 /// Signed bamboo entry, encoded as a hexadecimal string.
-#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
+#[derive(Scalar, Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
+#[graphql(name = "EntryEncoded")]
 pub struct EncodedEntryScalar(EncodedEntry);
 
-#[Scalar(name = "EncodedEntry")]
-impl ScalarType for EncodedEntryScalar {
-    fn parse(value: Value) -> InputValueResult<Self> {
+impl ScalarValue for EncodedEntryScalar {
+    fn from_value(value: Value) -> Result<Self> {
         match &value {
             Value::String(str_value) => {
                 let bytes =
-                    hex::decode(str_value).map_err(|_| anyhow!("invalid hex encoding in entry"))?;
+                    hex::decode(str_value).map_err(|e| anyhow!(e.to_string()))?;
                 Ok(EncodedEntryScalar(EncodedEntry::from_bytes(&bytes)))
             }
-            _ => Err(InputValueError::expected_type(value)),
+            _ => Err(Error::new(format!("Expected a valid encoded entry, found: {value}"))),
         }
     }
