+++ aquadoggo/src/lib.rs
@@ -47,7 +47,6 @@ mod bus;
 mod config;
 mod context;
 mod db;
-mod dynamic_graphql;
 mod graphql;
 mod http;
 mod manager;
