+++ aquadoggo/src/replication/service.rs
@@ -19,7 +19,6 @@ use tokio::task;
 use crate::bus::{ServiceMessage, ServiceSender};
 use crate::context::Context;
 use crate::db::types::StorageEntry;
-use crate::graphql::replication::client;
 use crate::manager::{ServiceReadySender, Shutdown};
 
 /// Replication service polling other nodes frequently to ask them about new entries from a defined
@@ -43,59 +42,7 @@ pub async fn replication_service(
                 debug!("Starting replication with remote peers");
             }
 
-            // Ask every remote peer about latest entries of log ids and authors
-            for remote_peer in remote_peers.clone().iter() {
-                for public_key_to_replicate in public_keys_to_replicate.clone().iter() {
-                    let public_key = public_key_to_replicate.public_key();
-                    let log_ids = public_key_to_replicate.log_ids().clone();
-
-                    for log_id in log_ids {
-                        // Get the latest sequence number we have for this log and public key
-                        let latest_seq_num =
-                            get_latest_seq_num(&context, &log_id, public_key).await;
-                        debug!(
-                            "Latest entry sequence number of {} and {}: {:?}",
-                            log_id.as_u64(),
-                            public_key,
-                            latest_seq_num
-                        );
-
-                        // Make our replication request to the remote peer
-                        let response = client::entries_newer_than_seq_num(
-                            remote_peer,
-                            &log_id,
-                            public_key,
-                            latest_seq_num.as_ref(),
-                        )
-                        .await;
-
-                        match response {
-                            Ok(entries) => {
-                                debug!(
-                                    "Received {} new entries from peer {}",
-                                    entries.len(),
-                                    remote_peer
-                                );
-
-                                if let Err(err) = verify_entries(&entries, &context).await {
-                                    warn!("Couldn't verify entries: {}", err);
-                                    continue;
-                                }
-
-                                insert_new_entries(&entries, &context, tx.clone())
-                                    .await
-                                    .unwrap_or_else(|err| error!("{:?}", err));
-                            }
-                            Err(err) => {
-                                warn!(
-                                    "Replication request to peer {} failed: {}",
-                                    remote_peer, err
-                                );
-                            }
-                        }
-                    }
-                }
-            }
+            // TODO: Replication goes here.....
 
             // Wait a couple of seconds before we attempt next replication requests
             tokio::time::sleep(connection_interval).await;
@@ -245,133 +192,4 @@ async fn get_latest_seq_num(
         .ok()
         .flatten()
         .map(|entry| *entry.seq_num())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::convert::TryInto;
-    use std::time::Duration;
-
-    use p2panda_rs::storage_provider::traits::EntryStore;
-    use p2panda_rs::test_utils::memory_store::helpers::{populate_store, PopulateStoreConfig};
-    use rstest::rstest;
-    use tokio::sync::{broadcast, oneshot};
-    use tokio::task;
-
-    use crate::context::Context;
-    use crate::http::http_service;
-    use crate::replication::ReplicationConfiguration;
-    use crate::test_utils::shutdown_handle;
-    use crate::test_utils::{
-        doggo_fields, doggo_schema, test_runner_with_manager, TestNodeManager,
-    };
-    use crate::Configuration;
-
-    use super::replication_service;
-
-    // @TODO: This will be replaced with using `ctor` in this PR:
-    // https://github.com/p2panda/aquadoggo/pull/166
-    fn init_env_logger() {
-        let _ = env_logger::builder().is_test(true).try_init();
-    }
-    #[rstest]
-    fn full_replication() {
-        test_runner_with_manager(|manager: TestNodeManager| async move {
-            init_env_logger();
-
-            // Build and populate Billie's database
-            let billie = manager.create("sqlite::memory:").await;
-            billie.context.schema_provider.update(doggo_schema()).await;
-            let populate_db_config = PopulateStoreConfig {
-                no_of_entries: 1,
-                no_of_logs: 1,
-                no_of_public_keys: 1,
-                with_delete: false,
-                schema: doggo_schema(),
-                create_operation_fields: doggo_fields(),
-                update_operation_fields: doggo_fields(),
-            };
-            let (key_pairs, _) = populate_store(&billie.context.store, &populate_db_config).await;
-
-            // Launch HTTP service of Billie
-            let (tx, _rx) = broadcast::channel(120);
-            let tx_billie = tx.clone();
-            let shutdown_billie = shutdown_handle();
-            let context_billie = Context::new(
-                billie.context.store.clone(),
-                Configuration {
-                    http_port: 3022,
-                    ..Configuration::default()
-                },
-                billie.context.schema_provider.clone(),
-            );
-            let (tx_ready, rx_ready) = oneshot::channel::<()>();
-
-            let http_server_billie = task::spawn(async {
-                http_service(context_billie, shutdown_billie, tx_billie, tx_ready)
-                    .await
-                    .unwrap();
-            });
-
-            if rx_ready.await.is_err() {
-                panic!("Service dropped");
-            }
-
-            // Our test database helper already populated the database for us. We retreive the
-            // public keys here of the authors who created these test data entries
-            let public_key = key_pairs.first().unwrap().public_key();
-            let log_ids: Vec<u64> = vec![0];
-            let endpoint: String = "http://localhost:3022/graphql".into();
-
-            // Construct database and context for Ada
-            let config_ada = Configuration {
-                replication: ReplicationConfiguration {
-                    public_keys_to_replicate: vec![(public_key.to_string(), log_ids)
-                        .try_into()
-                        .unwrap()],
-                    remote_peers: vec![endpoint],
-                    ..ReplicationConfiguration::default()
-                },
-                ..Configuration::default()
-            };
-            let ada = manager.create("sqlite::memory:").await;
-            ada.context.schema_provider.update(doggo_schema()).await;
-            let context_ada = Context::new(
-                ada.context.store.clone(),
-                config_ada,
-                // We want ada to have the same schema in her schema provider as billie.
-                ada.context.schema_provider.clone(),
-            );
-            let tx_ada = tx.clone();
-            let shutdown_ada = shutdown_handle();
-            let (tx_ready, rx_ready) = oneshot::channel::<()>();
-
-            // Ada starts replication service to get data from Billies GraphQL API
-            let replication_service_ada = task::spawn(async {
-                replication_service(context_ada, shutdown_ada, tx_ada, tx_ready)
-                    .await
-                    .unwrap();
-            });
-
-            if rx_ready.await.is_err() {
-                panic!("Service dropped");
-            }
-
-            // Wait a little bit for replication to take place
-            tokio::time::sleep(Duration::from_millis(500)).await;
-
-            // Make sure the services did not stop
-            assert!(!http_server_billie.is_finished());
-            assert!(!replication_service_ada.is_finished());
-
-            // Check the entry arrived into Ada's database
-            let entries = ada
-                .context
-                .store
-                .get_entries_by_schema(doggo_schema().id())
-                .await
-                .unwrap();
-            assert_eq!(entries.len(), 1);
-        })
-    }
 }
